backendold/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.12.0
python-dotenv==1.0.0
qdrant-client==1.14.3
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
numpy==1.24.3
python-multipart==0.0.6
httpx==0.25.2